Refactor connection ODP path management
The `updateConnection` API function now expects the `odpPath` to be an
array of ODP point IDs directly. The UI components have been updated to
manage the `odpPath` as an array of IDs.

This includes adjusting Svelte-Select usage for multi-select with
`bind:value` and initializing connection `odpPath`s as IDs. A filter for
`null` values was also added before API submission.